[lainonlife] Disable nginx proxy_temp files.
This was causing a very significant disk space leak (see the Grafana
graph!) where nginx was writing the entire audio stream of each
channel to disk.

This was pretty hard to track down, as the proxy_temp files got
deleted: they were however held open by the worker process.  This
meant that `du` did not help.  I eventually remembered that `iotop`
existed, and once I saw that an nginx worker was continuously writing,
a quick look at `lsof` was enough to track this down.