[innsmouth] Fix LetsEncrypt certs for barrucadu.{com,uk}
With the global redirect, LE was generating certs for barrucadu.co.uk,
rather than the original domain.

I suspect the only reason this worked at all is because there's a
common shared well-known directory across all the hosts - so LE made
the request, got 301ed to a different domain, got the correct token,
and so issued a cert for the new domain.

Surprisingly, there *is* a `location` block in the config for the
well-known directory, but I guess the global redirect (corresponding
to a `return 302 ...;` in the server block) took priority.  Work
around that be putting the redirect in a location block for /.