[nyarlathotep] Prepare for new root device
I've ordered an SSD to use as the boot/root volume, so I can use the
current HDDs entirely for NAS purposes.  The current ZFS pool will be
cleaned up and used entirely for NAS purposes.  I've moved everything
that needs keeping to /mnt/data, so the paths in the current and new
systems are the same.

When the new SSD arrives I'll regenerate the hardware config; rename
the ZFS pool to "data"; delete everything outside of /mnt/data and
move that to the root of the pool; and reinstall NixOS.