[nyarlathotep] Fix pihole destination graph
The "destination" metric seem to use two formats: "name|ip" and just
"ip".  The ones without a name don't seem to be changing, so just
disregard those ones.

There's a similar problem with the pihole top sources, but the entries
with just an IP *are* changing, so ideally for that I'll find some way
to merge based on IP but also keep the hostname.