[nyarlathotep] Automatically tag music
I tried to use systemd-path for this, but it doesn't support the
pattern needed for my podcasts, which is:

    /mnt/nas/music/Podcasts/*/in/*.mp3

It doesn't seem to support multiple `*`s in the same pattern, as
replacing either one of them with a concrete filename works.  I don't
want to list all the podcasts I listen to in my NixOS config, so a
regular unit which runs `inotifywait -r` in a loop will do.

systemd-path does work for converting albums, as there I only need the
simple pattern:

    /mnt/nas/music/to_convert/in/*