[dreamlands] Fix event streaming for Concourse
Requests to the events endpoint, which is of type text/event-stream,
would fail, and so prevent live events from showing in the Concourse
UI.

The Caddy docs say:

> flush_interval is a duration value that defines how often Caddy
> should flush the buffered response body to the client. Set to -1 to
> disable buffering. It is set to -1 automatically for requests that
> have a text/event-stream response or for HTTP/2 requests where the
> Content-Length is unspecified.

So it should have just worked, as the mime type matches, but it
doesn't.  Explicitly setting the flush_interval solves the problem.