[nyarlathotep] Use .lan domains for http
Firefox on Windows has stopped liking "http://nyarlathotep/" - it
redirects me to "https://www.nyarlathotep.com", without even making a
request to the right server.  All the *.lan domains work fine, so
after failing to solve this problem on the Firefox end I'm giving up
and using "nyarlathotep.lan" for HTTP.

The server is still accessible as just "nyarlathotep" for SSH, ping,
etc.  It's just Firefox that is being a pain.