[nyarlathotep] Make runway metrcis work with time ranges
This change drops the first expenses[£] value, which isn't great... but
it's only really a problem if that first value is a significant chunk of
the expenses in that time range.  That's only likely to be a problem in
very small ranges, and in very small ranges the metric isn't meaningful
anyway.