[carcosa] Add grafana dashboard
This dashboard has already paid off, as it was while making this that
I noticed the unusually high disk usage by Concourse (#12).

I think this is a pretty good dashboard which could potentially become
a template for other machines, as it only uses the node and docker
exporters.  It does, however, have knowledge of how mountpoints are
allocated to partitions, and what alerts are appropriate.  So it's not
fully generic.