Tighten up docker unit dependencies
`After`, `BindsTo`, and `Requires` all do slightly different things:

`After` imposes an ordering constraint.  If unit `X` has `After=Y` and
both `X` and `Y` are started, then `X` will be delayed until `Y` has
finished starting.

So we want `After=docker.service`, as `docker` and `docker-compose`
won't work if the daemon is not up yet.

`BindsTo` and `Requires` impose dependencies.  If a unit `X` has
`BindsTo=Y` or `Requires=Y`, and `X` is started then `Y` will be
started if it not already running.

Combining `Requires=Y` (or `BindsTo=Y`) and `After=Y` means that
starting `X` will

- Start `Y`
- Wait for `Y` to finish starting
- Then start `X`

If there is no `After`, then instead `X` and `Y` will be started in
parallel.

`BindsTo` also makes `X` stop if `Y` stops.  This is particularly
useful for units which bind-mount the docker socket (as the
prometheus-docker-exporter does), as if the docker daemon restarts
then the old socket is invalid.