Fix Concourse cgroup issue
After a recent reboot, Concourse stopped working.  The culprit seems
to be a cgroup issue as described in the linked GitHub issue (see
diff), so it's likely a systemd or kernel upgrade which broke it.

Since everything else is otherwise working fine, this is better than
rolling back to an older generation.