Upgrade Concourse to 7.4
Unfortunately, there's a DNS resolution issue with this: sometimes,
when using the default resolver (which I believe comes from docker, as
it's on a 172 IP), names fail to resolve, which generally makes a job
fail.

So, skip the local DNS resolution entirely and use 1.1.1.1, falling
back to 8.8.8.8 if it's down.  So far this looks to solve the issue.