[nyarlathotep] Use variables to simplify dashboard & add FIRE deltas
Adds `$average_daily_expense`, `$fire_annual_factor`, and
`$fire_invested` to simplify the FIRE and runway calculations.

The `$fire_annual_factor` is 1 / SWR: the number of years expenses you
need saved to be able to safely withdraw one year's expenses
annually.  This is 25, as I'm using the 4% rule.

Not sure why, but the generated dashboard JSON downgrades a bunch of
components.  There isn't an obvious visual impact of this...

This commit also fixes a couple of unintentional changes introduced by
d50631d: changing the default time period to 2 years, and expanding
the market prices row.