[nyarlathotep] Add AAW, PAW, and FIRE thresholds to assets graph
According to The Millionaire Next Door, a person's expected net
worth (sans inheritance) is:

```
pre-tax income sans inheritance * age / 10
```

People with under half of that are UAWs (underaccumulators of wealth),
people with more than double that are PAWs (prodigious accumulators of
wealth), and everyone else is an AAW (average accumulator of wealth).

So, add these thresholds to the assets graph.

First, I've worked out average daily income sans gifts, which I'm
using as a stand-in for inheritance here.  They're kind of the same
thing: money I've received for nothing, rather than earned from work
or an investment.

```
(-(last("income[£]") - first("income[£]") + last("income:gift[£]") - first("income:gift[£]")) / count("income[£]"))
```

Remember that income is negative: hence the signs!

The threshold for being an AAW is, then:

```
($average_daily_income * 365 * $age / 10) / 2 + OFFSET
```

Where the OFFSET:

1. Shifts the threshold UP to counter-act any gifts:

    ```
    - last("income:gift[£]")
    ```

2. Shifts the threshold UP to counter-act any liabilities:

    ```
    - last("liabilities[£]")
    ```

3. Shifts the threshold DOWN for any not-real-liabilities:

    ```
    + last("liabilities:loan:slc[£]")
    ```

Remember that liabilities, like income, are negative: hence the signs!

The `/ 2` is because that's the upper limit for a UAW, therefore
anything above that is a AAW.

The PAW formula is similar: but the `/ 2` is a `* 2`.

I've also added the FIRE threshold.  An interesting observation is
that FIRE is above PAW, so all financially independent people are
prodigious accumulators of wealth.  Makes sense.  Though, since the
AAW / PAW thresholds discount gifts, I could in principle be given
enough money to reach FIRE, but that wouldn't make me a better
"accumulator of wealth": which also makes sense, as a gift isn't
something I've "accumulated".

Since these thresholds are all far above my current level of assets,
I've set an upper limit for the Y axis.  I expect increasing that
limit as I accumulate wealth to feel very satisfying.

I'll need to tweak this every year, to increase the `$age` variable.