[nyarlathotep] Adjust budget goals
- Drop general emergency fund: I already have 3+ months expenses in
cash, and a credit card, and general-purpose floats.  It's hard to
imagine what would come up which couldn't be solved by reallocating
money from those.

- Reduce float to £250/bank: if I need more than that, I need to
adjust the budget.

- Reduce move fund to £1000: this is likely still too high, given I'm
not likely to move for a while, but it could happen that I'm suddenly
served an eviction notice and need to move out.

I'll invest as a lump sum the excess money that was in those
categories.