Add firewall module with fail2ban and ip blocking
There are a lot of connection attempts to my whitelisted Minecraft
server.  I occasionally check the Minecraft logs, and it's kind of
annoying having to mentally filter through these.  Since they come
from a handful of IPs, it's nice to have a way to block them.

This is done with a new iptables chain, put before the nixos-fw chain,
which just either drops a connection if it's on the naughty list, or
returns to the INPUT chain.

I only really need this on carcosa, because that's where the affected
service is, but there's no reason not to make this a module which all
hosts pull in.

And, since I'm adding one sort of blocking, I thought I may as well
also turn on fail2ban, which will (by default) block SSH connection
attempts.