Bind Pi Hole to server IP, not 0.0.0.0
If there are other docker containers running on the same host, and the
host resolves DNS via the Pi Hole, then those other containers are
unable to resolve DNS.  Specifically, they get a failure in name
resolution due to the response coming back from an unexpected IP.

The solution is to bind to the external IP only, not to all
interfaces.  Note that this now rules out the `onlyLocalDNS` option.