[nyarlathotep] Rewrite hledger scripts to use promscale & add to repo
promscale[1] is a prometheus-compatible database which has an API to
insert old data, and to delete old data.  So I can use it to
reimplement my existing scripts without significantly changing the
drop-and-bulk-import workflow.

But, since PromQL is much more powerful than the Influx v1 query
language, I actually need to import far less data to get the same
functionality!  So the scripts are simpler - and I took this
opportunity to rewrite them in Python and move to this repository.
I'll update the old hledger-scripts repo to refer here, and then
archive it.

This also updates all the dashboard panels to the latest widget
versions.

[1] https://github.com/timescale/promscale