Switch to a flake-based configuration
This commit drops `configuration.nix` and the `host` symlink in favour
of a somewhat repetitive `flake.nix`.  Flakes seem to be the direction
the Nix / NixOS ecosystem is moving in, I'll have to make this change
eventually, so why not now?  Plus, this lets me benefit from the stuff
already set up for flakes, like sops-nix.

I still have auto-updates enabled, so the nixpkgs hash in the
`flake.lock` will quickly fall out of date, but that's just a bit of
noise in future PRs, not a huge issue.

THIS IS IMPURE CURRENTLY!  That's because of the way I'm handling
secrets.  I plan to switch to sops-nix (or something similar) to solve
this issue.