Add a service file for Foundry
Foundry's got a few different bits of config now, so move it all to a
new file (even though it's only used by carcosa).

This relies on the Foundry files being in the right place.  Since
Foundry is nonfree software which has to be downloaded from the online
store after purchase, that seems unavoidable.