[nyarlathotep] Import DNS config
Previously, this was done by unmanaged (but backed-up) files in
`/persist`.  But now that the default config is taken straight from
the resolved source (5ab0007), only the custom config needs to be
managed.

Custom zones I've added to the secrets (though it's more "private"
than "secret") and the StevenBlack blocklist I just fetch from the
source.  Yes, this does impose a new external dependency on this repo,
but the chances of it disappearing seem pretty slim, and I can just
remove it if that does happen.  The alternative was committing ~6MB of
hosts file, which felt a bit excessive.