Switch to standard Pleroma module
The standard Pleroma module is good enough, and means I don't need to
build my own containers from the develop branch - which isn't a
supported approach anyway and frequently breaks.

I need to use the Pleroma from nixpkgs-unstable because my current image
is based on 2.5, but nixpkgs only has 2.4, which isn't compatible due to
database changes.

The database is still dockerised, so it's not subject to the whims of
`system.stateVersion`.

I've also changed the backup script to use the custom pgdump format,
because that can be restored in parallel.  A test run only took 90
minutes (plus another 9 minutes to run a `VACUUM ANALYZE`), which is
much better than the 3 *days* it took when I last had to restore a SQL
dump.