[nyarlathotep] Implement new way for remote sync
Now that carcosa has a user (and scripts) to receive the bookdb &
bookmarks syncs, remove the current approach of doing everything as my
user & grepping systemd units for paths & ports.

It's not ideal that the remote-sync user needs to be in the docker
group, but this is just one more piece of motivation to run bookdb
outside of docker.