Run bookdb outside of docker
Natively packaging this application isn't very hard.  It does make the
sync and backup scripts a little more complex, as they now need to
handle retrieving files owned by the bookdb user, but that's not too
tricky.

This also makes deploying updates to nyarlathotep a little easier: I
don't need to build a docker image, push it to a local repo, and restart
the service any more.  I just bump the version in the package, same as
bookmarks, resolved, and prometheus-awair-exporter.

Needing to expose the Elasticsearch instance to localhost is a bit
unfortunate, however.  Maybe I'll be able to do some systemd network
namespacing magic and lock that down at some point.

I also changed the bookdb port to be a random one in the high end of the
user ports range which is unlikely to clash with anything else.

This also lets me remove the remote-sync users from the `docker` group,
and remove the concourse-deploy-robot from carcosa.  The docker
registries on carcosa and nyarlathotep are still needed for now, though.