[nyarlathotep] Offset sync from backup schedule
These occasionally collide right now.  Usually that makes the backup
fail (since the collision tends to be with the bookdb cover files
changing while the backup `tar` is running, which throws an error), but
if it happened during the elasticsearch index sync I could end up
backing up incomplete data.

Another option would be adding some sort of lockfile which prevents both
running at the same time, but this is simpler.