Make pods a first-class citizen for OCI containers
To migrate to podman I want my oci-containers lib to do all the heavy
lifting.  Ideally, I'll be able to switch back and forth between podman
and docker on a single host without breaking anything.

As part of that, add a 'pod' abstraction which groups containers
together.  All containers in a pod share the same `volumeSubDir`.

With docker, pods are faked by putting the containers onto the same
network.  But with podman they're true pods.