Add script to generate options documentation
This necessitated a little refactoring, splitting the options
declarations out of the config so that I don't need to somehow evaluate
all the standard nixos modules just to generate documentation for my
stuff.

For now, the script treats missing documentation as a warning, not an
error.  Once everything is documented, this will change.