[nyarlathotep] Implement workaround to forecast finances
VictoriaMetrics only allows data up to 2 days in the future, which is a
pain for forecasting (am I cursed to be unable to find a timeseries
database that meets all my wants at once?).  So implement a work-around:
an optional env var specifying an integral number of years to shift data
back in time by.

Then if I want to, say, forecast up to the end of the year I run this
with `YEAR_OFFSET=1` and see that as up to the end of *last* year.